Refactored FilePartitionShims
