+++ sql-plugin/src/main/spark350/scala/org/apache/spark/sql/execution/rapids/shims/FilePartitionShims.scala
@@ -36,53 +36,4 @@ object FilePartitionShims {
       }
     }
   }
-
-  def splitFiles(sparkSession: SparkSession,
-      hadoopConf: Configuration,
-      selectedPartitions: Array[PartitionDirectory],
-      maxSplitBytes: Long): Seq[PartitionedFile] = {
-
-    def canBeSplit(filePath: Path, hadoopConf: Configuration): Boolean = {
-      // Checks if file at path `filePath` can be split.
-      // Uncompressed Hive Text files may be split. GZIP compressed files are not.
-      // Note: This method works on a Path, and cannot take a `FileStatus`.
-      //       partition.files is an Array[FileStatus] on vanilla Apache Spark,
-      //       but an Array[SerializableFileStatus] on Databricks.
-      val codec = new CompressionCodecFactory(hadoopConf).getCodec(filePath)
-      codec == null || codec.isInstanceOf[SplittableCompressionCodec]
-    }
-
-    selectedPartitions.flatMap { partition =>
-      partition.files.flatMap { f =>
-        PartitionedFileUtil.splitFiles(
-          sparkSession,
-          f,
-          isSplitable = canBeSplit(f.getPath, hadoopConf),
-          maxSplitBytes,
-          partition.values
-        )
-      }.sortBy(_.length)(implicitly[Ordering[Long]].reverse)
-    }
-  }
-
-  def splitFiles(selectedPartitions: Array[PartitionDirectory],
-      relation: HadoopFsRelation,
-      maxSplitBytes: Long): Array[PartitionedFile] = {
-
-    selectedPartitions.flatMap { partition =>
-      partition.files.flatMap { file =>
-        // getPath() is very expensive so we only want to call it once in this block:
-        val filePath = file.getPath
-        val isSplitable = relation.fileFormat.isSplitable(
-          relation.sparkSession, relation.options, filePath)
-        PartitionedFileUtil.splitFiles(
-          sparkSession = relation.sparkSession,
-          file = file,
-          isSplitable = isSplitable,
-          maxSplitBytes = maxSplitBytes,
-          partitionValues = partition.values
-        )
-      }
-    }.sortBy(_.length)(implicitly[Ordering[Long]].reverse)
-  }
 }
